#26 bug修复: cython问题修复及与python结果对比

+++ codes/buffett/cython/zdf/node.pxd
@@ -1,6 +1,6 @@
 cdef class Node:
     cdef public:
-        float val
+        double val
         int size
         Node left, right
